requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
bitly-api==0.3
celery==3.1.16
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-angular==0.7.9
django-bootstrap3==4.11.0
django-bower==5.0.1
django-jenkins==0.16.3
django-nose==1.2
django-toolbelt==0.0.1
gunicorn==19.1.1
httplib2==0.9
kombu==3.0.23
mox==0.5.3
nose==1.3.4
psycopg2==2.5.4
pytz==2014.7
pyzmq==14.4.0
requests==2.4.3
selenium==2.44.0
six==1.8.0
static3==0.5.1
twilio==3.6.10
wsgiref==0.1.2
